+++ Tests/SheetMusicTests/EditingTests/EditReplayWebGoldenTests.swift
@@ -195,7 +195,7 @@
             case let .removeTuplet(at):
                 self.init(op: "removeTuplet")
                 self.at = ElementPath(at)
-            case .composite, .insertMeasure, .deleteMeasure, .addPart, .removePart, .movePart,
+            case .composite, .insertMeasure, .deleteMeasure, .addPart, .removePart, .movePart, .setPartNames,
                  .setKeySignature, .removeKeySignature, .setTimeSignature, .removeTimeSignature,
                  .setRehearsalMark, .removeRehearsalMark,
                  .createVoice, .splitRest, .setNoteHead, .setDrumsetEntry: